refactor initialization logic
initializers used to be evaluated as mop, individual by individual, this
is correct only in cases in which we do not need contextual info.
However for initial values this is not often the case as we normally
want a good distributed sample of the search space.

This commit separates the initialization logic to be a separate function,
this accepts an initializer object with all the needed parameters to
initialize any population.